Safe models- Lili CQ
This is the 3D model of the safe. The following repository includes a .3dm (editable rhino file), and interactive stl (untextured but allows the user to rotate it and look at it), and JPEG images of the model with textures applied using Fusion 360.

Why are we doing this? For our class, this modeling is necessary to aid our group in thinking in terms of the dimensions of the safe for building a robot for it.